Add certificate authentication on non-Windows platforms
Fix certificate authentication on non-Windows platforms due to the missing
cert store on Linux and a known issue in the PowerShell core commands which do not
support certificate files prior to this week. The new certificate
authentication supported by PowerGRR requires the PowerShell patch from
@markekraus. See issue #8 and README.

To handle certificate authentication on non-Windows platforms, a new
config option for the client certificate path was added
(GRRClientCertFilePath). It's now possible to use a certificate file
directly instead of using the Windows certificate store.
Get-PfxCertificate is used to read the certificate.

The certificate handling was tested on Windows with native PowerShell
v5 and on Windows and Ubuntu 16.04 with PowerShell 6.0 and @markekraus'
fork.

Resolves #8.