Fix an issue with the basic auth on non-Windows platforms
The credential parameter in the WebCmdlets in the OpenSource
implementation of PowerShell do not support basic auth implicitly.
Therefore, the credential handling was changed to explicitly use basic
auth headers in the web requests instead of the generic "credential"
parameter.

This allows using the same mechanism on Windows and on non-Windows
platforms. The authentication was tested on Windows with PowerShell v5,
Windows with PowerShell v6 and Ubuntu with PowerShell v6.

Fixes #11.